// link2rail/projects/ekp/src/app/empty-wagon/api/generated/model/demand-type.ts
/**
 * Empty wagon API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { NHM } from './nhm';
import { DemandWagonType } from './demand-wagon-type';
import { TransitRailwayUndertaking } from './transit-railway-undertaking';
import { CommercialLocation } from './commercial-location';


export interface DemandType { 
    /**
     * Maximum amount of wagons that can be ordered of this demand type.
     */
    maxOrderAmount?: number;
    /**
     * States whether a manual assessment of the request must occur in any case.
     */
    assessment?: boolean;
    shipperAuthorization?: boolean;
    commentCustomer?: string;
    referenceNumber?: string;
    demandWagonType: DemandWagonType;
    nhm?: NHM;
    loadRunLocation?: CommercialLocation;
    transitRailwayUndertaking?: TransitRailwayUndertaking;
}