// link2rail/projects/ekp/src/app/empty-wagon/api/generated/model/template.ts
/**
 * Empty wagon API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CustomerInformation } from './customer-information';
import { TemplateDemand } from './template-demand';


export interface Template { 
    /**
     * Name of template.
     */
    templateName: string;
    /**
     * Name of template, defined by customer.
     */
    customerTemplateName?: string;
    orderer: CustomerInformation;
    shipper?: CustomerInformation;
    demand: Array<TemplateDemand>;
}